Make ACQUIRE_ALL optional
